+++ src/utils/testUtils.ts
@@ -1,6 +1,6 @@
 import { default as chalk } from 'chalk';
 
-import { HttpResponse, ProcessorContext, TestFunction, TestResult, testSymbols } from '../models';
+import { HttpRegion, HttpResponse, ProcessorContext, TestFunction, TestResult, testSymbols } from '../models';
 import {
   assertHasNoResponseBody,
   assertHasResponseBody,
@@ -13,38 +13,52 @@ import {
 import { isError, parseError } from './errorUtils';
 import { isHttpResponse } from './requestUtils';
 
+export function testFactoryAsync({
+  httpRegion,
+  scriptConsole,
+}: ProcessorContext): (message: string, testMethod: () => Promise<void>) => Promise<void> {
+  return async function test(message: string, testMethod: () => Promise<void> | void): Promise<void> {
+    const testResult: TestResult = {
+      message,
+      result: true,
+    };
+    if (typeof testMethod === 'function') {
+      try {
+        await testMethod();
+      } catch (err) {
+        setErrorInTestResult(testResult, err);
+      }
+    }
+    addTestResultToHttpRegion(httpRegion, testResult, scriptConsole);
+  };
+}
+
+function setErrorInTestResult(testResult: TestResult, err: unknown) {
+  testResult.result = false;
+  if (isError(err)) {
+    testResult.error = parseError(err);
+  } else {
+    testResult.error = {
+      displayMessage: `${err}`,
+      error: new Error(`${err}`),
+    };
+  }
+}
+
 export function testFactory({ httpRegion, scriptConsole, variables }: ProcessorContext): TestFunction {
   const testFunction = function test(message: string, testMethod: () => void): void {
     const testResult: TestResult = {
       message,
       result: true,
     };
-    if (!httpRegion.testResults) {
-      httpRegion.testResults = [];
-    }
     if (typeof testMethod === 'function') {
       try {
         testMethod();
       } catch (err) {
-        testResult.result = false;
-        if (isError(err)) {
-          testResult.error = parseError(err);
-        } else {
-          testResult.error = {
-            displayMessage: `${err}`,
-            error: new Error(`${err}`),
-          };
-        }
+        setErrorInTestResult(testResult, err);
       }
     }
-    httpRegion.testResults.push(testResult);
-
-    scriptConsole?.logTest?.(
-      testResult.result,
-      testResult.result
-        ? chalk`{green ${testSymbols.ok} ${testResult.message || 'Test passed'}}`
-        : chalk`{red ${testSymbols.error} ${testResult.message || 'Test failed'} (${testResult.error?.displayMessage})}`
-    );
+    addTestResultToHttpRegion(httpRegion, testResult, scriptConsole);
   };
 
   function getHttpResponse(): HttpResponse | undefined {
@@ -100,4 +114,21 @@ export function testFactory({ httpRegion, scriptConsole, variables }: ProcessorC
     }
   };
   return testFunction;
+}
+function addTestResultToHttpRegion(
+  httpRegion: HttpRegion,
+  testResult: TestResult,
+  scriptConsole: import('/home/andreas/github/httpyac/httpyac/src/models/logHandler').ConsoleLogHandler | undefined
+) {
+  if (!httpRegion.testResults) {
+    httpRegion.testResults = [];
+  }
+  httpRegion.testResults.push(testResult);
+
+  scriptConsole?.logTest?.(
+    testResult.result,
+    testResult.result
+      ? chalk`{green ${testSymbols.ok} ${testResult.message || 'Test passed'}}`
+      : chalk`{red ${testSymbols.error} ${testResult.message || 'Test failed'} (${testResult.error?.displayMessage})}`
+  );
 }