+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ### Fix
 
 - allow import of `.httpyac.cjs` or `.httpyac.config.cjs` file
+- Memory Leak in grpc client, because grpc client is not closed (#231)
 
 ## [6.8.0] (2023-10-15)